Dowload Thermo Hela PRTC data

Downloads on of multiple Thermo Hela/PRTC data files.

getThermoHelaPRTC(src, destdir = ".")

Arguments

src
The name of the file to be downloaded. If missing, a vector of possible filenames is returned. If "all", all files are downloaded. Alternatively, a pattern can be used to grep the files from the output getThermoHelaPRTC() the files to be downloaded.

destdir
Destination directory. Default is ".".

Value

Invisibly return the path of the downloaded files.

See also

downloadData

Examples

getThermoHelaPRTC()

#> [1] "design.txt"                                 
#> [2] "swissprot_human_canonical_19_09_12.fasta.gz"
#> [3] "Thermo_Hela_PRTC_1.mgf.gz"                  
#> [4] "Thermo_Hela_PRTC_2.mgf.gz"                  
#> [5] "Thermo_Hela_PRTC_3.mgf.gz"                  

getThermoHelaPRTC("design")
not_run({
  getThermoHelaPRTC("all")
})