Proteomics and MS biocView packages

Searches for all the packages with the "Proteomics" (software), "MassSpectrometry" (software) and "MassSepctrometryData" (data) packages and return their names, titles and versions as a data.frame. The (unexported but documented) underlying functions are RforProteomics:::getPackagesInBiocView (to find relevant package) and RforProteomics:::packageDF (data.frame formatting).

proteomicsPackages(biocv, cache = FALSE)

massSpectrometryPackages(biocv, cache = FALSE)

massSpectrometryDataPackages(biocv, cache = FALSE)

Arguments

biocv
A character with the Bioconductor version to search for relevant packages. If missing, the running version is used.

cache
A logical indicating whether cached package information should be used. Default is FALSE. All except development versions are up-to-date.

Value

A data.frame with the respective package names, titles and versions.

Examples

head(pp <- proteomicsPackages("3.0"))

#> Warning: Dropping unknown biocViews terms:
#> CopyNumberVariants

#> Warning: Dropping unknown biocViews terms:
#> NoViewProvided

#> Warning: Dropping unknown biocViews terms:
#> ConnectTools, HighThroughputSequencing

#> Warning: Dropping unknown biocViews terms:
#> Bioinformatics

#> Warning: Dropping unknown biocViews terms:
#> HighThroughputSequencing

#> Warning: Dropping unknown biocViews terms:
#> Cancer

#> Warning: Dropping unknown biocViews terms:
#> Statistics, Bioinformatics

#> Warning: Dropping unknown biocViews terms:
#> MultipleComparisons

#> Warning: Dropping unknown biocViews terms:
#> CopyNumber

#> Warning: Dropping unknown biocViews terms:
#> Bioinformatics, Enrichment

#> Warning: Dropping unknown biocViews terms:
#> SNPsAndGeneticVariability

#> Warning: Dropping unknown biocViews terms:
#> NoViewProvided

#> Warning: Dropping unknown biocViews terms:
#> Bioinformatics

#> Warning: Dropping unknown biocViews terms:
#> Clustering.

#> Warning: Dropping unknown biocViews terms:
#> HighThroughputSequencing

#> Warning: Dropping unknown biocViews terms:
#> Bioinformatics, Cancer, FlowCytData, StemCells, DensityGating

#> Warning: Dropping unknown biocViews terms:
#> Gating

#> Warning: Dropping unknown biocViews terms:
#> GraphsAndNetworks

#> Warning: Dropping unknown biocViews terms:
#> NoViewProvided

#> Warning: Dropping unknown biocViews terms:
#> HighThroughputSequencing

#> Warning: Dropping unknown biocViews terms:
#> NoViewProvided

#> Warning: Dropping unknown biocViews terms:
#> Bioinformatics, MultipleComparisons

#> Warning: Dropping unknown biocViews terms:
#> Cancer

#> Warning: Dropping unknown biocViews terms:
#> Bioinformatics, Statistics, MultipleComparisons

#> Warning: Dropping unknown biocViews terms:
#> AMDIS, GCMS

#> Warning: Dropping unknown biocViews terms:
#> GraphsAndNetwork

#> Warning: Dropping unknown biocViews terms:
#> Bioinformatics

#> Warning: Dropping unknown biocViews terms:
#> HighThroughputSequencingData

#> Warning: Dropping unknown biocViews terms:
#> GraphsAndNetwork

#> Warning: Dropping unknown biocViews terms:
#> GO 

#> Warning: Dropping unknown biocViews terms:
#> NoViewProvided

#> Warning: Dropping unknown biocViews terms:
#> Infastructure

#> Warning: Dropping unknown biocViews terms:
#> NoViewProvided

#> Warning: Dropping unknown biocViews terms:
#> HTA2.0

#> Warning: Dropping unknown biocViews terms:
#> Bioinformatics, Visualisation

#> Warning: Dropping unknown biocViews terms:
#> Bioinformatics

#> Warning: Dropping unknown biocViews terms:
#> Bioinformatics

#> Warning: Dropping unknown biocViews terms:
#> HighThroughputSequencing

#> Warning: Dropping unknown biocViews terms:
#> NetworkAnalysis, GraphsAndNetworks

#> Warning: Dropping unknown biocViews terms:
#> NetworkAnalysis

#> Warning: Dropping unknown biocViews terms:
#> Cancer, StemCells, HIV

#> Warning: Dropping unknown biocViews terms:
#> Statistics

#> Warning: Dropping unknown biocViews terms:
#> Bioinformatics

#> Warning: Dropping unknown biocViews terms:
#> HighThroughputSequencing

#>             Package
#> ASEB           ASEB
#> bioassayR bioassayR
#> BRAIN         BRAIN
#> CellNOptR CellNOptR
#> ChemmineR ChemmineR
#> cisPath     cisPath
#>                                                                                                                   Title
#> ASEB                                                                                    Predict Acetylated Lysine Sites
#> bioassayR                                                                            R library for Bioactivity analysis
#> BRAIN                                                Baffling Recursive Algorithm for Isotope distributioN calculations
#> CellNOptR Training of boolean logic models of signalling networks using prior knowledge networks and perturbation data.
#> ChemmineR                                                                                 Cheminformatics Toolkit for R
#> cisPath                                       Visualization and management of the protein-protein interaction networks.
#>           Version
#> ASEB       1.10.0
#> bioassayR   1.4.3
#> BRAIN      1.12.0
#> CellNOptR  1.12.0
#> ChemmineR  2.18.1
#> cisPath     1.6.4

ppc <- proteomicsPackages("3.0", cache = TRUE)
all.equal(pp, ppc)

#> [1] "Component “Version”: 18 string mismatches"