PXD000001 example MSnSet

In this TMT 6-plex experiment, four exogenous proteins were spiked into an equimolar Erwinia carotovora lysate with varying proportions in each channel of quantitation; yeast enolase (ENO) at 10:5:2.5:1:2.5:10, bovine serum albumin (BSA) at 1:2.5:5:10:5:1, rabbit glycogen phosphorylase (PHO) at 2:2:2:2:1:1 and bovin cytochrome C (CYT) at 1:1:1:1:1:2. Proteins were then digested, differentially labelled with TMT reagents, fractionated by reverse phase nanoflow UPLC (nanoACQUITY, Waters), and analysed on an LTQ Orbitrap Velos mass spectrometer (Thermo Scientific). Files in multiple format will be used to illustrate the input/output capabilities that are available to the proteomics audience. The companion package provides dedicated functions to directly download the data.

The data has been downloaded from the ProteomeXchange repository and imported into R as illustrated in the example. It is of class MSnSet. See also the MSnbase-demo vignette for more details.

data("qnt")

Format

An instance of class MSnSet

References

Laurent Gatto (2014). RforProteomcs: Companion package to the 'Using R and Bioconductor for proteomics data analysis' publicationR package version 1.3.1.

Gatto L, Christoforou A. Using R and Bioconductor for proteomics data analysis. Biochim Biophys Acta. 2013 May 18. doi:pii: S1570-9639(13)00186-6. 10.1016/j.bbapap.2013.04.032. [Epub ahead of print] PubMed PMID: 23692960.

Examples

if (FALSE) {
    library("rpx")
    px1 <- PXDataset("PXD000001")
    mztab <- pxget(px1, "PXD000001_mztab.txt")
    library("MSnbase")
    qnt <- readMzTabData(mztab, what = "PEP")
    sampleNames(qnt) <- reporterNames(TMT6)
    qnt$conditions <- factor(c("A", "A", "B", "B", "B", "A"))
    qnt <- filterNA(qnt)

    selA <- qnt$conditions == "A"

    fData(qnt)$log2FC <-
        log(rowMeans(exprs(qnt)[, selA]), 2) -
            log(rowMeans(exprs(qnt)[, !selA]), 2)
    fData(qnt)$baseMean <- log(rowMeans(exprs(qnt)), 10)
}

library("RforProteomics")
library("MSnbase")
data(qnt)
class(qnt)

#> [1] "MSnSet"
#> attr(,"package")
#> [1] "MSnbase"

head(exprs(qnt))

#>    TMT6.126 TMT6.127 TMT6.128 TMT6.129 TMT6.130 TMT6.131
#> 2  10630132 11238708 12424917 10997763  9928972 10398534
#> 5  11105690 12403253 13160903 12229367 11061660 10131218
#> 6   1183431  1322371  1599088  1243715  1306602  1159064
#> 8   5384958  5508454  6883086  6136023  5626680  5213771
#> 10 18033537 17926487 21052620 19810368 17381162 17268329
#> 11  9873585 10299931 11142071 10258214  9664315  9518271

head(fData(qnt))

#>     sequence accession     unit_id unique database database_version
#> 2    DGVSVAR   ECA0625 PRIDE_22134   null  Erwinia                1
#> 5     NVVLDK   ECA0625 PRIDE_22134   null  Erwinia                1
#> 6  VEDALHATR   ECA0625 PRIDE_22134   null  Erwinia                1
#> 8  LAGGVAVIK   ECA0625 PRIDE_22134   null  Erwinia                1
#> 10  LIAEAMEK   ECA0625 PRIDE_22134   null  Erwinia                1
#> 11 SFGAPTITK   ECA0625 PRIDE_22134   null  Erwinia                1
#>              search_engine                      search_engine_score reliability
#> 2  [MS,MS:1001207,Mascot,] [PRIDE,PRIDE:0000069,Mascot score,26.92]        null
#> 5  [MS,MS:1001207,Mascot,] [PRIDE,PRIDE:0000069,Mascot score,22.66]        null
#> 6  [MS,MS:1001207,Mascot,]  [PRIDE,PRIDE:0000069,Mascot score,41.9]        null
#> 8  [MS,MS:1001207,Mascot,] [PRIDE,PRIDE:0000069,Mascot score,48.99]        null
#> 10 [MS,MS:1001207,Mascot,] [PRIDE,PRIDE:0000069,Mascot score,38.04]        null
#> 11 [MS,MS:1001207,Mascot,] [PRIDE,PRIDE:0000069,Mascot score,39.36]        null
#>              modifications retention_time charge mass_to_charge  uri
#> 2              0-MOD:01720           null   null           null null
#> 5  0-MOD:01720,6-MOD:01720           null   null           null null
#> 6              0-MOD:01720           null   null           null null
#> 8  0-MOD:01720,9-MOD:01720           null   null           null null
#> 10 0-MOD:01720,8-MOD:01720           null   null           null null
#> 11 0-MOD:01720,9-MOD:01720           null   null           null null
#>                 spectra_ref peptide_abundance_stdev_sub[1]
#> 2   ms_file[1]:spectrum=466                           null
#> 5  ms_file[1]:spectrum=1604                           null
#> 6  ms_file[1]:spectrum=2065                           null
#> 8  ms_file[1]:spectrum=2325                           null
#> 10 ms_file[1]:spectrum=2712                           null
#> 11 ms_file[1]:spectrum=2775                           null
#>    peptide_abundance_std_error_sub[1] peptide_abundance_stdev_sub[2]
#> 2                                null                           null
#> 5                                null                           null
#> 6                                null                           null
#> 8                                null                           null
#> 10                               null                           null
#> 11                               null                           null
#>    peptide_abundance_std_error_sub[2] peptide_abundance_stdev_sub[3]
#> 2                                null                           null
#> 5                                null                           null
#> 6                                null                           null
#> 8                                null                           null
#> 10                               null                           null
#> 11                               null                           null
#>    peptide_abundance_std_error_sub[3] peptide_abundance_stdev_sub[4]
#> 2                                null                           null
#> 5                                null                           null
#> 6                                null                           null
#> 8                                null                           null
#> 10                               null                           null
#> 11                               null                           null
#>    peptide_abundance_std_error_sub[4] peptide_abundance_stdev_sub[5]
#> 2                                null                           null
#> 5                                null                           null
#> 6                                null                           null
#> 8                                null                           null
#> 10                               null                           null
#> 11                               null                           null
#>    peptide_abundance_std_error_sub[5] peptide_abundance_stdev_sub[6]
#> 2                                null                           null
#> 5                                null                           null
#> 6                                null                           null
#> 8                                null                           null
#> 10                               null                           null
#> 11                               null                           null
#>    peptide_abundance_std_error_sub[6]      log2FC baseMean
#> 2                                null -0.04768204 23.38265
#> 5                                null -0.11581055 23.47779
#> 6                                null -0.17914403 20.31272
#> 8                                null -0.21114556 22.46567
#> 10                               null -0.12991824 24.14715
#> 11                               null -0.06520753 23.27157